+++ R/datadoc_cd-info.R
@@ -3,10 +3,11 @@
 #'
 #' Some of the most consequential variables to include in MRP are at the
 #' district-level. We include one such data for congressional districts. All data
-#' is collected by Daily Kos.
+#' is collected by Daily Kos. `cd_info_2018` is data on 2018 boundaries, `cd_info_2016`
+#' uses 2016 boundaries (but with place descriptions currently at 2016).
 #'
 #' @format `cd_info_2018` is a dataframe with the `r nrow(cd_info_2018)` Congressional
-#'  Districts, one row per cd.
+#'  Districts with 2018 boundaries, one row per cd.
 #'  \describe{
 #'    \item{year}{The year for the district line. A congressional district's
 #'    actual geography can change year to year, and significantly so in different
@@ -16,12 +17,13 @@
 #'    coding of \code{cd}: a two-letter abbreviation for the state followed by
 #'    a dash, and the district number padded with zeros to the left to be of length
 #'    2. At-large districts like Delaware are given a "-01" for the district number.}
-#'    \item{pct_trump, pct_romney, pct_mccain}{The voteshare of presidential candidates
+#'    \item{pct_trump, pct_romney, pct_mccain}{The two-party voteshare of Republican presidential candidates
 #'    in that district for the given year. E.g. the \code{pct_mccain} data when \code{cd_year == 2018}
-#'    represents the percent of the vote by McCain in 2008 for that district _under 2018 lines._}
-#'    \item{dailykos_name}{The unique descriptive name for the district given by Daily Kos.
+#'    represents the percent of the vote by McCain in 2008 for that district _under 2018 lines._
+#'    The Trump value is for 2016, not 2020.}
+#'    \item{dailykos_name}{The unique descriptive name for the district code in 2018 given by Daily Kos.
 #'    See Source for full citation.}
-#'    \item{largest_place}{The largest place in the district given by Daily Kos. Multiple districts may
+#'    \item{largest_place}{The largest place in the district code in 2018 given by Daily Kos. Multiple districts may
 #'    have the largest place.}
 #'  }
 #'
@@ -40,6 +42,10 @@
 #' head(cd_info_2018)
 #' head(elec_NY)
 "cd_info_2018"
+
+
+#' @rdname cd_info_2018
+"cd_info_2016"
 
 
 #' @rdname cd_info_2018