+++ data-raw/create_dist-level-samp.R
@@ -2,15 +2,25 @@
 # This loads from an external repo. Contact Shiro Kuriwaki for more
 
 library(tidyverse)
+library(fs)
 
-full_cd <- read_rds("~/Dropbox/CCES_representation/data/output/averages/by-cd_info.Rds")
+
+projdir <- "~/Dropbox/CCES_representation/data/output/"
 
 
-cd_info_2018 <- full_cd %>%
+cd_info_2018 <- read_rds(path(projdir, "averages/by-cd-2018_info.Rds")) %>%
   mutate(year = 2018) %>%
   select(year, cd,
          dailykos_name = descrip, largest_place = place,
          pct_trump, pct_romney, pct_mccain)
 
+cd_info_2016 <- read_rds(path(projdir, "averages/by-cd-2016_info.Rds")) %>%
+  mutate(year = 2016) %>%
+  select(year, cd,
+         dailykos_name = descrip_18,
+         largest_place = place_18,
+         pct_trump, pct_romney, pct_mccain)
 
+
+usethis::use_data(cd_info_2016, overwrite = TRUE)
 usethis::use_data(cd_info_2018, overwrite = TRUE)